Make Time comparison with Integer/String like Ruby
Previously, both Strings and Integers could be compared with a Time due
to Time's compare_with_coercion coercing the Time instate to a DateTime
before comparing. This does not align with Ruby's implementation of Time
comparison, which will raise when a Time is compared with a String or
Integer.

This commit fixes the mismatch by only casting Time to a DateTime if the
"other" acts like a Date.

This partially reverts 7872cc992bac76e976334d0c20649d69aad5652e, which
made the to_datetime conversion happen always instead of only when
"other" acts like a Date.